fix: structure of manage table
