+++ app/[locale]/admin/manage/page.tsx
@@ -1,13 +1,13 @@
-import VideosTable from "@/components/manage-vedio-table";
+import VideosTable from "@/components/manage-video-table";
 import { fetchMediaData } from "@/helpers/api/media.request";
 
 const Manage = async () => {
   const media = await fetchMediaData();
 
   return (
-    <div>
+    <>
       <VideosTable media={media} />
-    </div>
+    </>
   );
 };
