+++ app/[locale]/admin/page.tsx
@@ -1,9 +1,7 @@
-import { useTranslations } from "next-intl";
 import React from "react";
 
 const Admin = () => {
-  const t = useTranslations("admin");
-  return <div></div>;
+  return <></>;
 };
 
 export default Admin;
